POC update.
Functional Parametric EQ. Q settings are hidden from main screen and need to be updated by the CMD box. i.e. \b0q 2 or b4q 1
There is also work to be done to get the EQ variables filled with the saved settings upon startup. 
This little bug is annoying but eh what can ya do on V1. 
This release is POC and expects the user to set all bands during this session. of course the values will be saved, but they do not load to the eq routines on start yet..