updated user_settings.ini loader
Here are some changes to the loader that handles user_settings.ini . para_eq.c was quick to check the file and cause a faulting error. 
This fix will ensure that If the file is not there we will copy default_settings.ini to user_settings.ini and then reload as new.  This behavior typically will only occur on the first run after installation