FT8 behavior; logbook; ADIF export (#289)
* Ensure that we never send an empty R report

It seems the usual sequence is supposed to be
ft8_message_tokenize():
	strcpy(signal_strength, p);
ft8_on_start_qso():
    field_set("SENT", signal_strength);
ft8_process():
	report_send = field_str("SENT");
ft8_on_signal_report():
	sprintf(reply_message, "%s %s R%s", call, mycall, report_send);

But what if ft8_on_start_qso() didn't get called? I've seen cases where
I send out a useless packet like "A1XYZ LB2JK R"
If sbitx receives such a packet, it treats it as R+00; but maybe most
software will see it as invalid. I doubt such QSOs get logged.
So make sure that if "SENT" is not set, we set it to the signal strength
of the last packet (so that it's visible in the GTK SENT field before
we actually start sending the packet).

* Use ftx_message_encode_[non]std; gen audio @ pitch right before xmit

Whenever we are starting a QSO, pull call_to from the CALL field,
and extra from my grid.

When auto-replying (executing each state in the state machine), likewise
each of the 3 fields gets populated more directly, rather than
generating message text first and then having ftx_message_encode() pull
it apart again.

Separate sbitx_ftx_msg_audio() from the encode functions, and call it
just-in-time in ft8_start_tx(). Before, it was already spending the CPU
cycles to regenerate audio, but now we do it for the purpose of using
the current TX_PITCH each time. This way it's possible to adjust the
pitch at any time, to get around observed signals, or because of
guessing that if the other station didn't respond, maybe it would make
sense to transmit on a different audio frequency. Before, it was
necessary to hit ESC and then restart the call or the reply, just to use
a different pitch: that's not always possible, depending on how much
traffic has been logged to the console in the meantime.

Don't switch to transmit until/unless there is some audio to transmit
(avoid the tick-tick-tick).

* Add console_extract_semantic()

Since the console remembers the meanings of ranges of text, we can
extract them later.

Also fix a warning in modems.c

* Tapping console line sets X11 selection and calls new ft8_call()

Extract the callsign to call, and the other available data (grid, signal
strength, pitch, time) from the console semantics, rather than
re-parsing the whole line. I.e. we know which callsign we are going to
call because ft8lib told us which range of characters is the "de" call,
not because we just picked the second word in the message and hoped for
the best. Messages like "CQ POTA X1ABC AA01" will not throw it off:
ft8lib knows the callsign is X1ABC, not POTA.

Also rule out calling myself.

ft8_call() depends on fields having been set, but that makes behavior
a bit more predictable: the fields can be checked before calling.
Not that the code is making any use of that yet...

The X11 selection is useful in case you want to paste the whole message
into some other program (a text editor for manual logging, to look up
the callsign on QRZ or something). That works great with remote X11 too.

* Clean up logbook.c

- use tabs (controversial, but they allow you a choice of indentation size in your editor)
- remove trailing whitespace
- add .clang-format for reformatting other files the same way
- add_to_list(): descriptive variable names

* logbook: allow up to 11 characters for callsigns, 11 for date; safer strings

10 is not enough: FT8 allows 11, and here we need one more byte for the
trailing null. A 10-character date string also needs a null terminator.
5 characters is enough for a 4-digit time.

Use snprintf and strncpy everywhere in logbook.c.

* Log edit dialog: spell the Exchange sent and received labels better

* Log tx power and swr; ADIF export; show power in 1/10 units; comment editing

I changed the data/create_db.sql script, but of course I didn't want to
start over with a fresh database; so I did

  $ sqlite3 ~/sbitx/data/sbitx.db
  sqlite> alter table logbook add column tx_power text default "";
  sqlite> alter table logbook add column vswr text default "";

That puts the columns in the "wrong" order: they come last, but I'd
rather have the comment field last.  Verify it like this:

  sqlite> .schema logbook

logbook_add() detects whether or not the new columns have actually
been added, and omits them if not.

There's no ADIF field defined for SWR, AFAICT.

SWR and power are not editable in the edit dialog so far, but at least
they show up in the list.

But the code should not depend on column order: i.e. it should not do a
query like 'select * from logbook ...' and then expect the columns to be
in a particular order, for display purposes or otherwise. And there is
a lot of room for improvement in export_adif() still.

* Remove per-field newlines from ADIF export

I don't see the point; WSJT-X doesn't do it that way, and it made it
harder to read, harder to handle with simple tools like grep.

* Use PATH_MAX instead of warning-comments (technical debt)

* Make the Logged: message less confusing

* Fix warnings about strptime and struct tm; remove duplicate include

* write /tmp/sbitx_result_rows.txt instead of result_rows.txt; no json

This is extremely temporary data, and should therefore be written to a
ramdisk rather than wasting SD writes on that.

Also get rid of unused "json" arrays.

* Provide interface for logbook querying and ADIF export

A caller of these functions can query the logbook and export ADIF
without doing sqlite-specific things.

In other branches of this code base, there is another way: not just
saving to a file on the SD card, but exporting over a network interface
(a virtual filesystem). That's not integrated on this branch, but we
might as well make the code easier to keep in sync.

* Add IOTA/SOTA/POTA fields to logbook; export to ADIF, omit empty fields

I'm not 100% sure whether MY_POTA_REF is right for POTA and simply IOTA
is right for IOTA; but MY_SOTA_REF works for uploading SOTA logs.

We now omit empty fields when exporting ADIF. Most QSOs aren't expected
to have the xota fields, and for that matter there was no point in
exporting empty comment fields either. The ID is also omitted:
it's not important for uploading logs.

* build script: add new columns to database schema, if necessary

* Log editor: allow lengthening fields; fix overflows

Often, one may want to edit the QSO comment afterwards. It didn't make
sense to limit the text length of each log entry field to the string
that was already stored. Crashes were possible too.

* FT8: log frequency + pitch; ADIF export deals with it

Logs from WSJT-X are like this. Otherwise the "pitch" info would be lost.

On export, detect whether each row is a new log entry (in hz) or an
old one (in khz).