mouse wheel; web UI QSOs; fixes (#290)
* Mouse wheel: adjust the hovered field, not the focused field

I don't want to have to click first: it makes ergonomics worse for no reason.
Now the focus is for the encoder knob, and the mouse wheel is independent.

* Refactor ff_cs()

This was needed for zbitx work; on this branch, it just makes the code
more maintainable.

* Strip <> brackets from callsign before replying

The brackets just mean the callsign is hashed. If I see a message like
  JH6URJ <CQ35MD> -13
and click on it, I expect to attempt to call CQ35MD (although it's a
bit premature: better to wait for him to finish his QSO). Keeping the
brackets otherwise results in an error "Cannot parse message!"

Another time I clicked on the callsign <OK1PHM/QRP> and got this error:
  *Error: field[#contact_callsign] can't be set to [<OK1PHM/QRP>], improper size.
The field is 11 characters max, and that callsign now fits without the
brackets (as it should: up to 11 characters are allowed for FT8 callsigns.)

* index.html: "error parsing"; fix FT8 TX output; remove dup fn

The web UI was not showing any outgoing FT8 messages, only received
messages.

FT8_stylize_message_item() was defined twice.
log("FT8_process_message: error parsing ..." was repeated too many places:
also rather confusing.

Finally it turned out the actual problem was in deco_to_span(),
which is doing too much and assuming too much, especially about how many
tokens and their positional meaning (an oft-repeated mistake in this
code base). TBH I don't understand why we don't just loop over the
tokens and output them. Anyway, a TX message might have tokens
"#H143830","TX","1210","~","#FCQ","#RLA/K7IHZ","#HJO59"
and an RX message might have
"#G143745","27","-12","675","~","#FCQ","#GLB2JK","#GJO59"
of course we don't have separate confidence and strength values in a
TX message, so the tilde is at a different index. So search for it.

The right list for me is logically "QSOs", not only received messages:
I want to see what I send, as well as the received messages of interest
(having clicked on something in the "Band Activity" list, for example).

* Log incoming/outgoing FT8 messages, with type (LOG_INFO)

i3.n3 format, along with the same text that we send to the console,
both for queued TX messages and for decoded RX messages.

* Handle errors better in sbitx_ft8_encode and ft8_tx

sbitx_ft8_encode professes to return the error from ftx_message_encode(),
not replace all errors with -1.  Make it so.

ft8_tx() returns early if sbitx_ft8_encode() fails. And use strncpy
as a drive-by fix.

* FTx: send RR73 rather than RRR

It seems to be more standard, and why omit the 73?
Sometimes I send RRR a lot of times and get no response; not sure
if that's confusing for a state machine in some other software.