FT4; time format; fix power/swr logging (#292)
* Log max power and swr rather than just the field values

By the time we make a log entry, the transmitter is off.
So we have to keep track of it separately. It's better to use the max;
otherwise I've observed that the measured power sometimes tapers off
for the final measurement while transmission is stopping.

* Get rid of is_ft4 arguments

It's just a matter of which mode we're using; and the ftx_message_t
type is the same for both. FT4 is just a different audio encoding.

* Get FT4 basically working

* Fix web UI for FT4

* Don't log the LDPC decode messages; count CRC errors

LDPC decode errors are abundant: 10's of them per cycle on a dummy load.
It's too verbose for my taste when debugging the FTx state machine but
it can be uncommented again if necessary.

Likewise, count the (abundant) CRC errors and log the total rather than
logging each one.

* modem_ft8: common time format; try to decide what will be sent; decruft

That is, if the user hits the CQ button, or selects a message to answer,
decide whether the message should be FT4 or FT8, whether to send in the
even or odd slot, and whether it's time to start right now.

For the xOTA/Alt modes coming later, ftx_would_send() will also decide
which message to send; that's coming in a later patch.

All time stamps are formatted with tenth-of-a-second resolution now, but
the tenth is omitted if zero (which we often expect in FT8 mode).

* FTx: fix clock rounding

Show the assumed packet start time rather than wallclock time.
Amends 63c0e45ef5506edf39337d8cb247cafbcae60ae9

* FTx: don't log queued transmission to console unless delayed

* Rename ft8_... to ftx_... in modem_ft8.c; get rid of ft8_setmode()

ft8_setmode() was always incomplete. We have two fields instead:
ft8_tx1st (which will soon be renamed) and ft8_auto.