CQ alt and xOTA modes; autoresponder (#294)
* Add xOTA fields to settings dialog and logbook; macros

If xota is set, that and xota_loc get logged. ADIF export is already
done. So, change the settings before your IOTA/SOTA/POTA trip, then
you can export a log containing the references/IDs.

The IOTA/SOTA/POTA macros are similar: the first button sends e.g.
"CQ SOTA MYCALL GRID"; the most-essential buttons are the first 5
(because zbitx only provides 5 macro buttons, so we might as well
get used to using mostly those); and the last button allows you
to enter something else in the CALL field and send CQ THAT ...
(for example CQ BOTA on a beach, or CQ nnn with a 3-digit number).

Selection of the xota mode, xota_loc and macro set are independent.
Setting SOTA mode does not select SOTA.mc, but that's easy already.

* Honor current_macro setting; load macro when changing via mouse wheel

* check for errors in macro_load; don't override macro setting in modem_poll()

If I switched to SOTA macros for FT8/FT4, I want to keep it that way on startup,
not have FT8 mode always deciding that I must use the one-and-only FT8 macro set.

* Implement FTx CQ_alt and xOTA modes by changing ft8_tx1st to ftx_cq

In FT8 or FT4 modes, "CQ alt" means send CQ only on the first slot of a
4-slot period: i.e. in FT8 mode, send CQ only at 0 seconds and skip the
30-second CQ. In FT4 mode, send CQ only at 0/30 sec.

There are various reasons:
- listen for activity in the even slot as well as the odd
- space out the "repeats" to CQ for a longer period of time
- reduce heat
- conserve batteries
- from a receiver viewpoint, less-frequent CQing is probably
  just as noticeable while generating less clutter

"xOTA" mode means send the xota_loc location (reference ID) in
between the CQ messages, as free text:
- SOTA: send e.g. "SO LA/BU-119"
- POTA: send "PO <park ref>"
- IOTA: send "IO <island ref>"
Let's hope all these references are really 10 characters or less,
since "xO " takes 3 of 13 possible free-text characters.

I haven't found evidence of an existing convention like this,
so let's start one. I think it would be best to send the
location/reference "somehow", and the UI was not making this easy.

* Revamp FTx auto-responder

- extract callsign from the span that ft8_lib designated:
  that is, if the message is "CQ DX A1BCD AA00", the callsign is
  A1BCD, not "DX"
- actually override the first choice with POTA/SOTA/QRP/P if exists.
  Before, such a CQ could then be overridden by an ordinary CQ.
- fix race condition when repeat = 1: don't call_wipe() until
  the current outgoing message finishes transmitting
- store discrete time, snr, pitch in decodes table, rather than
  re-parsing displaytext; set fields from those (thus, update the
  UI logger fields and pitch "needle")
- the FT8_AUTO ON/OFF setting is replaced with FTX_AUTO OFF/ANS/CQRESP
  to clarify that it's also for FT4, and to separate answering messages
  directed at my callsign (ANS) from responding also to CQs from other
  hams (CQRESP). The OFF setting is not very useful: you have to be
  alert and ready to use the macro keys; so ANS is the default. Having
  CQ-responder as the default was extreme, IMO.

The next step beyond this should be called ROBOT, if it's going to
also send CQs when it runs out of CQs to answer. This is not done.