fix bugs: xota, mouse behavior after settings, and logging (#298)
* Fix gridsquare ADIF export for FT4; improve comments

SRX/STX should be gridsquare for FT4 too, not only FT8.

* xOTA failsafe: don't log xOTA if the location ref hasn't been set

If user_settings.ini has "#xota=" with no value, the GUI might treat it
as POTA rather than NONE (that needs fixing). So I sometimes saw log
entries with POTA PEAK/PARK/ISLE: XOTA got to be POTA somehow, and xOTA
Loc was not set. If it has the default text, it's not set, and this is
not a POTA log entry.

* If user_settings.ini has an empty string for "#xota", use NONE

Otherwise we would get a "not permitted" error, and it would default to
POTA in the settings dialog.

Alternatively, I could have fixed this by changing set_field() so that
if the argument is empty, it chooses the first option from an option field.
But this would mean that if you manually issue the command "\xota" with
no argument, it would go back to NONE; and likewise for every other option.
I'm not sure yet if we want that. Also don't understand why this problem
doesn't come up with other option fields.

* Fix bug: don't adjust volume by mouse move after settings dismissed

It was a weird one at first: the volume changes for no apparent reason.
Well... ok, it changes for no apparent reason after having opened and
closed the settings dialog. Hmm.

It seems that because we press a mouse button to press the button to
dismiss the settings dialog, the main window also gets a press event.
(That seems wrong, but maybe GTK just does that for some reason.)
So mouse_down=true, temporarily. But actually on_mouse_move() receives
a GdkEventMotion event, which contains button state, so we can reset
mouse_down every time. The only reason we need mouse_down now is for the
long-press feature on the virtual keyboard (assuming that still works
and is useful). It's from 694caaf04a941781334214558ca42bc3da7344af
Anyway, the bug occurred because we have the feature to adjust fields
by dragging (from 27006ea31759ffc0435b587ee4e7bc5c3306eac1).

Likewise, on_window_state() gets a GdkEventWindowState event, not
GdkEventKey. And given that on_mouse_move() keeps mouse_down updated,
I don't see a reason to change it in on_window_state().