bug fixes (#302)
* Move define LOG_LEVEL before ft8_lib includes; improve log messages

The LOG macro is provided by ft8_lib (ft8/debug.h), so we need to define
the log level beforehand if we want to change it from the default.

- FT8 modem: reverse arrows in LOG messages

I keep getting confused, which way is transmit and which way is receive?
The antenna jack is on the right side of the radio physically and on the
schematic too, so I guess it's easier to remember if ">> CQ ..." means
the signal is going out to the right, and "<< A1BCD ..." means it's
received. "-> ..." means it's queued up for later transmission.

Let's keep track of the number of repeats remaining, too. (Too bad the
UI doesn't make it clear)

* FT8: don't answer a messsage with my callsign if his callsign is invalid

E.g. this can be triggered by cold-calling from another radio that is using
a non-standard call. If we haven't seen a CQ from him/her, we don't have the
hash in the hashtable. So if the incoming message looks to us like
MYCALL <...> AB01  it's not enough information to be able to answer.
There's no point in sending " MYCALL Rxxx".

* CQRESP mode: don't look for another CQ to answer if any message is queued

Assume that if ftx_tx_text isn't empty, we have a message there that we
are sending or plan to send. (Therefore, it's now important to empty it out
whenever we don't have anything to send.)

* CQRESP mode: format SENT (rst) field with explicit sign and 2 digits

It will get logged from there, and it's nice to always have this format in the logbook.

* Log Hz-accurate frequency for FT4

Log dial frequency + audio frequency. This was meant to be for both
FTx modes. Amends d7ab3ac64043eced2eb4d08065c42e23922e0d32