Add and use ftx_call_or_continue (#312)
do_console() no longer has responsibility to extract fields from
whatever line the user clicked on: now we just hand the whole line and
its spans array to ftx_call_or_continue(), which does all the work,
including reimplementing the FTx state machine. This seems like a nice
simplification; and it also makes it possible to resume a QSO that was
aborted or timed out. It's similar to ft8_process() (which is probably
obsolete), except that it's given the spans array so that it can jump
more directly to the state machine without reparsing the message.

Amends d7ab3ac64043eced2eb4d08065c42e23922e0d32 which had introduced a
bug: calling only ft8_call when you click on a message meant that you
couldn't get it to take the next step to continue a QSO by clicking.