style RR73 as plain text not grid; add terminate arg to ft8_abort() (#319)
* sbitx_ft8_decode(): ensure RR73 is not tagged as a grid

The suspenders to go with the 1a4a94fc12c31a210462ccc17526856b70c53547 belt
We don't want RR73 standing out in the same style as a grid, in the GTK
console or in the web UI; so just tag it with STYLE_LOG, and then expect
it to have that style in ftx_call_or_continue().

* Add terminate_qso argument to modem_abort

If the user hits esc, the logger fields are cleared, so it makes sense
to also clear ftx_tx_text in the FTx modem: that means at the next
chance, we will not send anything unless we are starting a new QSO; and
in CQRESP mode, we are free to reply to another CQ, because there is
nothing "queued" up to send.

But in some cases, we don't want to clear ftx_tx_text: e.g. if a 73 needs
to be sent, we should finish sending it; or if we were retrying until
the number of repeats are exhausted, keep doing that.