Fix buffer overflows; add ftx_rx_pitch field independent of rx_pitch (#321)
* Fix a buffer overflow in do_control_action(TNPWR)

atoi() can read too far if the string is not null-terminated; and
this command could also be given a number that is out of range for
the band_stack array; so there were at least 2 ways to crash.

==28314==ERROR: AddressSanitizer: stack-buffer-overflow on address ...
READ of size 5 at 0x0... thread T4
    #0 in atoi (/home/pi/prj/sbitx/sbitx+0xaee3c)
    #1 in do_control_action src/sbitx_gtk.c:9835
    #2 in set_field src/sbitx_gtk.c:1312
    #3 in field_set src/sbitx_gtk.c:1361
    #4 in apply_band_settings src/sbitx_gtk.c:4873
    #5 in set_operating_freq src/sbitx_gtk.c:4901
    #6 in tx_off src/sbitx_gtk.c:7483
    #7 in ftx_call_or_continue src/modem_ft8.c:1395
    #8 in sbitx_ft8_decode src/modem_ft8.c:889

* Add ftx_rx_pitch field to hold chosen station pitch in FTx modes

It needed to be separated from rx_pitch, because that is a user setting:
in CW mode, it's the audio frequency the user would prefer to hear, and
one doesn't expect it to change on its own. Whereas ftx_rx_pitch jumps
each time you choose an FT8/FT4 message to respond to, and controls
the position of the green "needle" in those modes.

* Fix another crash

	for (old_band = 0; old_band < max_bands; old_band++)
        if (whatever)
            break;
    int stack = band_stack[old_band].index;

Clearly old_band could end at max_bands if the loop didn't break,
and is then out of range for band_stack indexing.

==7117==ERROR: AddressSanitizer: global-buffer-overflow on address 0x0...
READ of size 4 at 0x0... thread T0
    #0 in change_band src/sbitx_gtk.c:9379
    #1 in do_band_dropdown src/sbitx_gtk.c:5992
    #2 in on_mouse_press src/sbitx_gtk.c:8015