FTx CQ country lookup (#332)
* Add clu submodule

This will be used to map callsigns and grids from FTx messages to
countries and distances.

* Start using clu to look up country, abbreviation, distance and azimuth

Only on CQ messages. It's no problem to look up every callsign, but
it's more work for the CPU and more clutter: maybe we don't need it.

* Add semantic spans for country, distance and azimuth

Could be different colors, but bright green looks nice enough,
and matches the color for CQ.

* Automatically update cty.dat into the data directory; use that one