control-W to close log window; logbook fixes (#334)
* Clean up logbook memory management

ASAN complains about memory leaks from this code, but at least the leaks
do not grow much when the logbook is repeatedly opened and closed.

It seems that GTK deletes widgets by default when the window containing
them is closed, so we don't need to add any free/destroy calls for this.

* Handle control-W to close the logbook window

* ADI export: strlen again after stripping hyphens from date (8 chars not 10)

I was catching up on QSLs. tqsl pointed out that e.g. <QSO_DATE:10>20251221 < ... was
too much. It's 8 characters not 10.