Deal with FTx lines that are too long to fit; logging fixes; maybe decode a bit more FT4 (#339)
* Add trailing newline to SWR console warning

When outputting a few characters to the console (as in CW mode),
a newline is not written, and the line continues until it wraps.
But a full-line console message should end with a newline.
If this is not done, the semantic spans from whatever comes next
(such as an FT8 message) might get applied to this unfinished line too,
which can be colorful but strange.

* Don't clear the logger after 73 is sent in manual mode

The intention with a75dc6de22e6e0a412e27303b2ce2dc334d2db32 was to delay
clearing it until after 73 finishes sending in any of the automated modes.
But in manual mode, we should not clear it at all. Manual means manual:
the user needs to take every step, including logging the QSO; so we shouldn't
interfere by clearing the fields.

In any automated mode, ftx_call_or_continue() logs the QSO when it sees
RR73 or RRR. But in manual mode, it returns early and does not do that;
so as a matter of symmetry, we shouldn't call_wipe() in manual mode either.

* Logbook: add ftx_rx_pitch to dial frequency, not rx_pitch

It was right until 7c45d64c3c3cbe86890671d125b5f977cd80061a

* Remove early return in ft8_rx()

The same idea was expressed two different ways: a bit below, we have
```
	if (slot_time < 500)
		ftx_rx_buff_index = 0;
```
I.e. start at the beginning of the sample buffer. We should _not_ return
early _without_ resetting ftx_rx_buff_index, in case it might miss something
next time it's called.

* Partially revert bigfont

This reverts most of commit ec3c812b0e9db932e7a2ea187eb6b96283487e10
so that it can be redone as a minimal patch.

* Restore bigfont without the reformatting

Here's the minimal patch with just the functionality, so we can actually
see the changes, just those lines in all their "glory".

This is not my change, and I don't approve of how it's done. But I need
to be able to rebase my other patches on top of something closer to what
it was before.

* Deal with long FTx lines on the console; style dist & azimuth

Don't wrap, don't truncate ahead of time, just avoid painting spans
that don't fit horizontally.

Thus, we can store more fields in each console line than what we can
afford to paint. We need console_extract_semantic() to always work.

And since font size can now be changed, we need to avoid wrapping
in FT8 mode. If you set the font too big, you just can't see the
whole message, let alone the country etc.: that's your bad choice.

Whether or not to send all the fields to the web UI is another matter.
Currently we do send them, so it's getting crowded there. But the
web UI could also truncate lines if someone wants to force it to
be narrower.

Distance and azimuth have unique styles: yellow like the grid, but
a little different. Smaller font for the web UI to help them fit.

* Fix the FTx state machine handling of RR73 and 73 yet again

If country lookup occurs on such messages,
strstr(line + line_len - 5, "RR73") will not work, because
the "RR73" is no longer that close to the end of the line.
It's more efficient not to search the whole line for it.
Whereas looking for RR73 in specific spans was also brittle,
as long as we don't have a special semantic for such things
as 73, RRR and RR73 (not even sure what to call those "words").
Anyway, the intention with doing country lookup when "my call"
is found was just to handle the case when somebody answers my
CQ with a grid. So let's do it only when there is actually a grid.