Add UI for FTx rules (#343)
* ftx_rules: add CRUD functions in preparation for an editing GUI

I wrote the function signatures, ChatGPT 5 mini implemented them,
and I wrote the doc comments and made minor corrections. We'll see.

* Add conversions ftx_rules_field <-> string

* Add a GUI window for editing FTx rules

ChatGPT wrote a lot of this, but it needed a good bit of work afterwards.
The prompt:

  Please implement ftx_rules_ui(). Available API for editing rules is in
  ftx_rules.h.

  What I have in mind is a window with a table view (probably GtkTreeView)
  in the upper area, with 3 columns: description and the two priority
  fields.  Clicking on any row there will populate some fields below the
  table view, according to the type of rule, and hide the irrelevant
  fields.  There are two kinds of rules: numeric rules (RULE_FIELD_SNR
  RULE_FIELD_DISTANCE or RULE_FIELD_AZIMUTH) and regex rules (the rest,
  which are applied to string fields in an FT8 or FT4 message).  Use
  spinboxes for the priority values, and for min and max if it's a numeric
  rule; use a single-line text field for the description, and another text
  field for the regex if it's that kind of rule.  Use a combobox to select
  the type of field to which the rule applies (the enum values in
  ftx_rules_field).

  If cq_resp_pri_adj and ans_pri_adj are both zero, the rule will be
  ignored at runtime: somehow you can indicate that in the table (use
  disabled text style or whatever).

  This will be built with a C compiler, not C++.

It wrote everything inside the ftx_rules_ui() function, including several
nested callback functions. Nested functions are supposed to work, as a
GNU C extension. But it was crashing, and gdb was not giving any clues;
so I had to restructure it with all the callbacks outside, and therefore
all the widget pointers have to be "global", but they are static to avoid
external linkage. It also made some parts more complicated than necessary:
for example it's a feature that all possible field names are sequential,
so it's easy to convert from a combobox index <-> enum ftx_rules_field
rather than doing unnecessary string comparisons.

test/rules is a way to test the rule editing feature without waiting for
the rest of the sbitx app to start up. It can be used as a standalone
rule-editing app.

* Rules UI: change priorities immediately; add button clears fields

You hit the add button first, fill in details of your new rule,
then save it. Whereas before, if you had selected an old rule,
everything would be populated: then it would be problematic to
change priorities immediately while the spinboxes are adjusted,
because that would affect the old rule, and then you'd save the
new rule afterwards and be confused.

Deleting and re-creating rules is still problematic though because
the database is auto-incrementing the IDs, and we store them in
memory as signed 8-bit. You can have max 128 rules then: there
needs to be a way to reuse old rule IDs when old rules are deleted.
Maybe ID is not a good ordering criterion either...

* Reuse IDs in ftx_rules; don't save unless needed

Since we cram rule ID into 8 bits, we need to avoid having "churn"
increase IDs beyond that range, in case the user does lots of editing.
Some will likely save after changing priorities on old rules too,
even though it's not needed; so don't delete and re-create it if
there are no other changes.